Added code to implement session management so that the expenses are manages as per the user logged in.
Added code to implement session management so that the expenses are manages as per the user logged in.

The commit is an initial commit for the implementation and in future it will have modifications such that the user is allowed to check their expenses, add new expenses as well as delete in a session.